replace with parameters helper function
